requirements.txt
PyQt6==6.6.1
requests==2.31.0
python-dotenv==1.0.1
sounddevice==0.4.6
numpy>=2.1.0
openai==1.13.3
